requirements.txt
aiogram
aiofiles
aiosqlite
redis
python-dotenv